Online Shopping prediction.
The data set was acquired from kaggle, the main aim of this project was to find out the intention of the visitors on an online shopping website. The Target column is revenue so i intended to find out if this particular intention vistors intention ended into a closed sale which in turn brings revenue.